+++ con-organizer/lib/hooks/UseSingleEvent.tsx
@@ -1,17 +1,14 @@
 import { useEffect, useState } from 'react';
 import { ConEvent } from '../../models/types';
-import { singleEvent$ } from '../observables/SingleEvent';
+import { singleEvent$ } from '../observable';
 
 export const useSingleEvents = (id: string) => {
     const [event, setEvent] = useState<ConEvent>();
     const [loading, setLoading] = useState<boolean>(true);
 
     useEffect(() => {
-        const eventObservable = singleEvent$(id).subscribe((snapshot) => {
-            if (snapshot.data()) {
-                setEvent({ ...(snapshot.data() as ConEvent), id: snapshot.id });
-                setLoading(false);
-            }
+        const eventObservable = singleEvent$(id).subscribe((event) => {
+            setEvent(event as ConEvent);
             setLoading(false);
         });
