--- con-organizer/lib/observables/AllEvents.ts
@@ -1,6 +1,0 @@
-import { collection } from 'firebase/firestore';
-import { collectionData } from 'rxfire/firestore';
-import { db } from '../firebase';
-
-export const eventsRef = collection(db, 'events');
-export const allEvents$ = collectionData(eventsRef, { idField: 'id' });
